+++ components/SummarizeDisplay.tsx
