--- components/audioFileTable/loading.tsx
@@ -1,4 +1,0 @@
-export default function Loading() {
-    // Or a custom loading skeleton component
-    return <p>Loading...</p>
-}
