+++ lib/summarize.ts
@@ -29,35 +29,6 @@ async function handleSummaryDatabase(fileId: number, summaryText: string) {
 }
 
 
-const endPoint = {
-    "translate": "https://api-inference.huggingface.co/models/t5-base",
-    "summarize": "https://api-inference.huggingface.co/models/google/flan-t5-base"
-}
-
-const summarizing = {
-    "casual": "Please revise this sentence to a less formal tone: 'I am going to the store.'",
-    "timeline": "Create a timeline summary of the following transcription. Highlight key points or topics discussed at different time intervals, ensuring that the essence of the conversation or presentation is captured succinctly. Include time stamps for easy reference."
-}
-export const summarize = async (formalSentence: string) => {
-    const prompt = `${summarizing.timeline}: '${formalSentence}'`;
-    const data = {
-        inputs: prompt,
-    }
-
-    const response = await fetch(
-        endPoint.summarize, // Replace with your target model
-        {
-            headers: { Authorization: `Bearer ${process.env.HF_ACCESS_TOKEN}` },
-            method: "POST",
-            body: JSON.stringify(data),
-        }
-    );
-    const result = await response.json();
-    const summary = result[0].generated_text;
-    console.log(summary);
-    return summary;
-}
-
 const ChatGptSummarize = async (formalSentence: string) => {
     const OpenAIendPoint = {
         "summarize": "https://api.openai.com/v1/chat/completions"
